map-marketingcloud-worker/requirements.txt
pycrypto
urllib3
Werkzeug==0.16.1
Flask==1.1.1
flask-restplus==0.13.0
requests==2.22.0
PyMySQL
python-dotenv
git+http://StashRO:BaRu3Res@git.bvops.net/scm/COM/kasasa-common-python.git#egg=kasasa_common[vault]
